backend/requirements.txt
fastapi==0.115.12
uvicorn==0.21.1
SQLAlchemy==2.0.40
mysql-connector-python==8.0.32
pydantic==2.11.4
pymysql
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
azure-storage-blob
python-dotenv
python-multipart
faiss-cpu
openai
pytest
httpx
sqlalchemy
alembic
pydantic-settings